+++ src/app/components/ErrorBoundary.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Component, ReactNode } from 'react';
 
 type Props = {
