+++ src/app/components/Search.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent, FormEvent, useContext, useEffect, useState } from 'react';
 import { SelectedItem } from '../constants/interfaces';
 import { Loader } from './Loader';
-import { useGetPageQuery, useGetPersonQuery } from '../../api/apiSlice';
-import { ThemeContext } from '../App';
+import { useGetPageQuery, useGetPersonQuery } from '../api/apiSlice';
+import { ThemeContext } from './App';
 import classNames from 'classnames';
 
 interface SearchProps {
