+++ src/app/page/ClientOnly.tsx
@@ -3,8 +3,9 @@
 import dynamic from 'next/dynamic';
 import { Provider } from 'react-redux';
 import { store } from '../store/store';
+import React from 'react';
 
-const App = dynamic(() => import('../App'), { ssr: false });
+const App = dynamic(() => import('../components/App'), { ssr: false });
 
 export function ClientOnly() {
   return (
