+++ src/app/page/page.tsx
@@ -1,0 +1,7 @@
+import { ClientOnly } from './ClientOnly';
+
+import React from 'react';
+
+export default function Page() {
+  return <ClientOnly />;
+}
