+++ src/app/reducers/selectedItemsSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { selectedState } from '../app/constants/interfaces';
+import { selectedState } from '../constants/interfaces';
 import { RootState } from '../store/store';
 
 const initialState: selectedState = {
